refactor(memory): complete RecallPipeline deep decomposition and inner class removal (#487)
- Remove ParallelScanEmitter and SequentialScanEmitter inner classes from RecallPipeline into package com.spectrayan.spector.memory.pipeline.scan.
- Move RRF BM25 candidate fusion logic into RecallCandidateGatherer.
- Move cognitive post-scoring and prospective reminder seeding into SalienceAndHabituationScorer.
- Ensure RecallPipeline delegates all candidate gathering, scoring, and graph expansion tasks to dedicated phase components.
